# fix(intl-messageformat-parser): no need to escape less than sign in messages