--- a/packages/intl-messageformat-parser/tests/benchmark.js
+++ b/packages/intl-messageformat-parser/tests/benchmark.js
@@ -9,21 +9,21 @@ const complexMsg =
   'female {' +
   '{num_guests, plural, offset:1 ' +
   '=0 {{host} does not give a party.}' +
-  '=1 {{host} invites {guest} to her party.}' +
-  '=2 {{host} invites {guest} and one other person to her party.}' +
-  'other {{host} invites {guest} and # other people to her party.}}}' +
+  '=1 {{host} invites <em>{guest}</em> to her party.}' +
+  '=2 {{host} invites <em>{guest}</em> and <em>one</em> other person to her party.}' +
+  'other {{host} invites <em>{guest}</em> and <em>#</em> other people to her party.}}}' +
   'male {' +
   '{num_guests, plural, offset:1 ' +
   '=0 {{host} does not give a party.}' +
-  '=1 {{host} invites {guest} to his party.}' +
-  '=2 {{host} invites {guest} and one other person to his party.}' +
-  'other {{host} invites {guest} and # other people to his party.}}}' +
+  '=1 {{host} invites <em>{guest}</em> to his party.}' +
+  '=2 {{host} invites <em>{guest}</em> and one other person to his party.}' +
+  'other {{host} invites <em>{guest}</em> and <em>#</em> other people to his party.}}}' +
   'other {' +
   '{num_guests, plural, offset:1 ' +
   '=0 {{host} does not give a party.}' +
-  '=1 {{host} invites {guest} to their party.}' +
-  '=2 {{host} invites {guest} and one other person to their party.}' +
-  'other {{host} invites {guest} and # other people to their party.}}}}';
+  '=1 {{host} invites <em>{guest}</em> to their party.}' +
+  '=2 {{host} invites <em>{guest}</em> and one other person to their party.}' +
+  'other {{host} invites <em>{guest}</em> and <em>#</em> other people to their party.}}}}';
 
 const normalMsg =
   '' +