Fixes #1845

Also adds benchmark script. Perf stats:

```
Baseline:

complex_msg AST length 2635
normal_msg AST length 400
simple_msg AST length 79
string_msg AST length 36
complex_msg x 4,273 ops/sec ±0.36% (93 runs sampled)
normal_msg x 34,153 ops/sec ±1.32% (93 runs sampled)
simple_msg x 165,777 ops/sec ±2.80% (95 runs sampled)
string_msg x 193,458 ops/sec ±0.30% (94 runs sampled

This branch:

complex_msg AST length 2599
normal_msg AST length 400
simple_msg AST length 79
string_msg AST length 36
complex_msg x 3,844 ops/sec ±1.62% (91 runs sampled)
normal_msg x 27,693 ops/sec ±6.56% (83 runs sampled)
simple_msg x 139,535 ops/sec ±4.36% (88 runs sampled)
string_msg x 180,446 ops/sec ±1.81% (88 runs sampled)
```

11% slower for complex messages.